Downloads/files (3)/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Config:
    SECRET_KEY = os.getenv("SECRET_KEY", "dev-secret-change-in-production")
    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URL", "sqlite:///guardlinkx.db")
    SQLALCHEMY_TRACK_MODIFICATIONS = False

    STRIPE_SECRET_KEY = os.getenv("STRIPE_SECRET_KEY", "")
    STRIPE_PUBLISHABLE_KEY = os.getenv("STRIPE_PUBLISHABLE_KEY", "")
    STRIPE_WEBHOOK_SECRET = os.getenv("STRIPE_WEBHOOK_SECRET", "")

    ADMIN_EMAIL = os.getenv("ADMIN_EMAIL", "admin@guardlinkx.com")
    ADMIN_PASSWORD = os.getenv("ADMIN_PASSWORD", "changeme")

    DOMAIN = os.getenv("DOMAIN", "califaxvpn.guardlinkx.com")
