Downloads/files (3)/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.0.7
stripe==11.4.1
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
email-validator==2.2.0
requests==2.32.3
maturin>=1.0,<2.0
